@@ -1,91 +1,102 @@
 {
 	"name": "pirate",
 	"version": "1.0.0",
+	"type": "module",
 	"private": false,
 	"scripts": {
 		"dev": "astro dev",
+		"dev:host": "astro dev --host 0.0.0.0",
+		"dev:netlify": "netlify dev",
 		"start": "astro dev",
-		"build": "./update-theme.sh && astro build",
-		"postbuild": "pagefind --site dist",
+		"build": "npm run update-theme && astro build && pagefind --site dist",
+		"update-theme": "bash ./update-theme.sh",
 		"preview": "astro preview",
 		"format": "prettier -w --cache . && prettier -w --cache **/*.astro",
 		"check": "astro check",
-		"clean": "rm -rf .data dist .astro .netlify node_modules"
+		"clean": "rm -rf .data dist .astro .netlify .vite node_modules",
+		"keystatic": "keystatic"
 	},
 	"dependencies": {
-		"@astrojs/markdoc": "^0.11.5",
-		"@astrojs/mdx": "3.1.5",
-		"@astrojs/netlify": "^5.5.4",
-		"@astrojs/react": "^3.6.3",
-		"@astrojs/rss": "4.0.7",
-		"@astrojs/sitemap": "3.1.6",
-		"@astrojs/tailwind": "5.1.0",
+		"@astrojs/markdoc": "^0.15.0",
+		"@astrojs/netlify": "^6.6.3",
+		"@astrojs/react": "^4.4.2",
+		"@astrojs/rss": "4.0.14",
+		"@astrojs/sitemap": "3.6.0",
+		"@astrojs/tailwind": "^6.0.2",
 		"@astrojs/vercel": "^7.8.2",
-		"@iconify-json/bi": "^1.2.2",
-		"@iconify-json/game-icons": "^1.2.1",
-		"@iconify-json/mdi": "^1.2.2",
-		"@iconify-json/simple-icons": "^1.2.16",
-		"@keystar/ui": "^0.7.14",
-		"@keystatic/astro": "^5.0.3",
-		"@keystatic/core": "^0.5.42",
+		"@iconify-json/bi": "^1.2.4",
+		"@iconify-json/game-icons": "^1.2.3",
+		"@iconify-json/mdi": "^1.2.3",
+		"@iconify-json/simple-icons": "^1.2.43",
+		"@keystar/ui": "^0.7.19",
+		"@keystatic/astro": "^5.0.6",
+		"@keystatic/core": "^0.5.48",
 		"@pagefind/default-ui": "^1.3.0",
 		"@reach/router": "^1.3.4",
 		"@tailwindcss/aspect-ratio": "^0.4.2",
-		"@tailwindcss/typography": "^0.5.15",
-		"@types/react": "18.3.5",
-		"@types/react-dom": "18.3.0",
-		"@vite-pwa/astro": "^0.4.3",
-		"astro": "4.16.7",
-		"astro-icon": "^1.1.4",
-		"autoprefixer": "^10.4.20",
-		"dotenv": "^16.4.7",
+		"@tailwindcss/typography": "^0.5.10",
+		"@types/react": "19.1.8",
+		"@types/react-dom": "19.1.6",
+		"@vite-pwa/astro": "^1.2.0",
+		"astro": "5.16.4",
+		"astro-icon": "^1.1.5",
+		"autoprefixer": "^10.4.17",
+		"dotenv": "^17.2.0",
 		"es-define-property": "^1.0.1",
 		"es-errors": "^1.3.0",
 		"fetch-blob": "^4.0.0",
 		"has-property-descriptors": "^1.0.2",
 		"lodash": "^4.17.21",
+		"lodash-es": "^4.17.21",
+		"marked": "^16.1.2",
 		"mdast-util-to-string": "^4.0.0",
 		"node-fetch": "^3.3.2",
 		"pagefind": "^1.3.0",
-		"postcss": "^8.4.49",
-		"postcss-html": "^1.7.0",
-		"postcss-import": "^16.1.0",
-		"qs": "^6.13.1",
-		"react": "18.3.1",
+		"patch-package": "^8.0.1",
+		"postcss": "^8.4.35",
+		"postcss-html": "^1.8.0",
+		"postcss-import": "^16.1.1",
+		"qs": "^6.14.0",
+		"react": "19.1.0",
 		"react-color": "^2.19.3",
 		"react-colorful": "^5.6.1",
-		"react-dom": "18.3.1",
-		"reading-time": "^1.5.0",
+		"react-dom": "19.1.0",
 		"rehype-external-links": "^3.0.0",
-		"remark-unwrap-images": "^4.0.1",
 		"rss-parser": "^3.13.0",
 		"set-function-length": "^1.2.2",
-		"sharp": "^0.33.5",
-		"stripe": "^17.5.0",
-		"tailwindcss": "^3.4.17",
-		"yet-another-react-lightbox": "^3.21.7"
+		"sharp": "^0.34.3",
+		"stripe": "^10.3.0",
+		"tailwindcss": "^3.4.1",
+		"workbox-window": "^7.3.0",
+		"yaml": "^2.8.0",
+		"yet-another-react-lightbox": "^3.25.0"
 	},
 	"devDependencies": {
 		"@astrojs/check": "^0.9.4",
+		"@eslint/markdown": "^7.5.0",
+		"@netlify/functions": "^5.1.0",
 		"@types/eslint": "^9.6.1",
-		"@types/lodash": "^4.17.13",
-		"@typescript-eslint/eslint-plugin": "^8.18.1",
-		"@typescript-eslint/parser": "^8.18.1",
-		"astro-eslint-parser": "^1.1.0",
-		"eslint": "^9.17.0",
-		"eslint-import-resolver-typescript": "^3.7.0",
+		"@types/lodash": "^4.17.20",
+		"@types/node": "^24.10.0",
+		"@typescript-eslint/eslint-plugin": "^8.37.0",
+		"@typescript-eslint/parser": "^8.37.0",
+		"astro-eslint-parser": "^1.2.2",
+		"eslint": "^9.31.0",
+		"eslint-import-resolver-typescript": "^3.10.1",
 		"eslint-plugin-astro": "^1.3.1",
 		"eslint-plugin-deprecation": "^3.0.0",
-		"eslint-plugin-import": "^2.31.0",
+		"eslint-plugin-import": "^2.32.0",
 		"eslint-plugin-jsx-a11y": "^6.10.2",
-		"eslint-plugin-markdown": "^5.1.0",
 		"eslint-plugin-perfectionist": "^3.9.1",
-		"eslint-plugin-regexp": "^2.7.0",
+		"eslint-plugin-regexp": "^2.9.0",
 		"eslint-plugin-typescript-sort-keys": "^3.3.0",
-		"prettier": "^3.4.2",
+		"prettier": "^3.6.2",
 		"prettier-plugin-astro": "0.14.1",
-		"prettier-plugin-tailwindcss": "^0.6.9",
-		"typescript": "^5.7.2"
+		"prettier-plugin-tailwindcss": "^0.5.12",
+		"typescript": "^5.8.3"
 	},
-	"packageManager": "pnpm@9.5.0+sha512.140036830124618d624a2187b50d04289d5a087f326c9edfc0ccd733d76c4f52c3a313d4fc148794a2a9d81553016004e6742e8cf850670268a7387fc220c903"
+	"packageManager": "pnpm@10.24.0",
+	"engines": {
+		"node": ">=18.0.0"
+	}
 }
