@@ -8,6 +8,16 @@ import "yet-another-react-lightbox/plugins/thumbnails.css";
 export default function GalleryLightbox({ images, showCaptions }) {
   const [open, setOpen] = useState(false);
   const [index, setIndex] = useState(0);
+  const [currentImages, setCurrentImages] = useState(images);
+
+  // Read auto-open setting from DOM
+  useEffect(() => {
+    const autoOpenElement = document.querySelector('[data-auto-open]');
+    const shouldAutoOpen = autoOpenElement?.getAttribute('data-auto-open') === 'true';
+    if (shouldAutoOpen) {
+      setTimeout(() => setOpen(true), 100);
+    }
+  }, []);
 
   const options = {
     settings: {
@@ -59,7 +69,7 @@ export default function GalleryLightbox({ images, showCaptions }) {
       thumbnailsIconColor: '#ffffff',
       thumbnailsOpacity: 0.4,
       thumbnailsPosition: 'bottom',
-      thumbnailsSize: ['100px', '80px']
+      thumbnailsSize: ['120px', '90px']
     },
     progressBar: {
       backgroundColor: '#f2f2f2',
@@ -70,21 +80,70 @@ export default function GalleryLightbox({ images, showCaptions }) {
   };
 
   useEffect(() => {
-    const imageElements = document.querySelectorAll('.post-card1 img');
-    imageElements.forEach((img, idx) => {
-      img.style.cursor = 'pointer';
-      img.addEventListener('click', () => {
-        setIndex(idx);
-        setOpen(true);
+    const setupClickListeners = () => {
+      const imageElements = document.querySelectorAll('.post-card1:not(.hidden) img');
+      
+      // Build array of visible images in order using their src
+      const visibleImages = [];
+      const srcToVisibleIndex = new Map();
+      
+      imageElements.forEach((img) => {
+        const imgSrc = img.getAttribute('data-src');
+        // Find this image in the full images array by matching src
+        const matchingImage = images.find(i => i.src === imgSrc);
+        if (matchingImage && !srcToVisibleIndex.has(imgSrc)) {
+          srcToVisibleIndex.set(imgSrc, visibleImages.length);
+          visibleImages.push(matchingImage);
+        }
       });
+      
+      // Update current images to only visible ones
+      if (visibleImages.length > 0) {
+        setCurrentImages(visibleImages);
+      }
+      
+      imageElements.forEach((img) => {
+        img.style.cursor = 'pointer';
+        
+        // Remove existing listener
+        const newImg = img.cloneNode(true);
+        img.parentNode.replaceChild(newImg, img);
+        
+        // Add new listener - find index in visible images by src
+        const imgSrc = newImg.getAttribute('data-src');
+        const visibleIndex = srcToVisibleIndex.get(imgSrc) || 0;
+        
+        newImg.addEventListener('click', () => {
+          setIndex(visibleIndex);
+          setOpen(true);
+        });
+      });
+    };
+    
+    setupClickListeners();
+    
+    // Re-setup listeners when images are filtered
+    const observer = new MutationObserver(() => {
+      setTimeout(setupClickListeners, 100);
     });
-  }, []);
+    
+    const galleryContainer = document.querySelector('[data-section-id*="gallery"] .homegallery');
+    if (galleryContainer) {
+      observer.observe(galleryContainer, { 
+        attributes: true, 
+        subtree: true, 
+        attributeFilter: ['class', 'style'] 
+      });
+    }
+    
+    return () => observer.disconnect();
+  }, [images]);
 
   return (
     <Lightbox
       open={open}
       close={() => setOpen(false)}
-      slides={images}
+      slides={currentImages}
       index={index}
       plugins={[Thumbnails]}
       {...options}
