@@ -9,12 +9,33 @@ function removeDupsAndLowerCase(array: string[]) {
 
 const postSchema = z.object({
   title: z.string(),
-  description: z.string().min(50).max(160),
-  publishDate: z.string().or(z.date()).transform((val) => new Date(val)),
+  description: z.string().min(10).max(160),
+  publishDate: z.string().or(z.date()).transform((val) => {
+    if (val instanceof Date) return val;
+    // For date-only strings (YYYY-MM-DD), treat as local date not UTC
+    if (typeof val === 'string' && /^\d{4}-\d{2}-\d{2}$/.test(val)) {
+      const [year, month, day] = val.split('-').map(Number);
+      return new Date(year, month - 1, day);
+    }
+    return new Date(val);
+  }),
+  updatedDate: z.string().or(z.date()).transform((val) => {
+    if (val instanceof Date) return val;
+    // For date-only strings (YYYY-MM-DD), treat as local date not UTC
+    if (typeof val === 'string' && /^\d{4}-\d{2}-\d{2}$/.test(val)) {
+      const [year, month, day] = val.split('-').map(Number);
+      return new Date(year, month - 1, day);
+    }
+    return new Date(val);
+  }).optional(),
   coverImage: z.object({
     src: z.string().optional(),
     alt: z.string().default(""),
   }).optional(),
+  overlayImage: z.string().optional(),
+  overlayImageAlt: z.string().optional(),
+  overlaySvg: z.string().optional(),
+  overlaySvgAlt: z.string().optional(),
   tags: z.array(z.string()).default([]).transform(removeDupsAndLowerCase),
   draft: z.boolean().default(false),
   order: z.object({
@@ -31,76 +52,125 @@ const postSchema = z.object({
       useCustomPlayer: z.boolean().optional(),
       mute: z.boolean().optional(),
       loop: z.boolean().optional(),
-      start: z.number().optional(),
-      end: z.number().optional(),
+      start: z.preprocess(
+        (val) => {
+          if (val === null || val === undefined || val === '' || (typeof val === 'number' && isNaN(val))) return undefined;
+          return typeof val === 'string' ? parseFloat(val) : val;
+        },
+        z.number().optional()
+      ),
+      end: z.preprocess(
+        (val) => {
+          if (val === null || val === undefined || val === '' || (typeof val === 'number' && isNaN(val))) return undefined;
+          return typeof val === 'string' ? parseFloat(val) : val;
+        },
+        z.number().optional()
+      ),
+      clickToLoad: z.boolean().optional(),
+      showMuteButton: z.boolean().optional(),
       videoOnly: z.boolean().optional(),
+      svgatorSync: z.boolean().optional(),
+      svgatorId: z.string().optional(),
+      interactiveMode: z.enum(['sync', 'independent', 'controller']).optional(),
+    }).optional()
+  }).optional(),
+  secondaryVideo: z.object({
+    discriminant: z.boolean(),
+    value: z.object({
+      url: z.string().optional(),
+      title: z.string().optional(),
+      controls: z.boolean().optional(),
+      useCustomPlayer: z.boolean().optional(),
+      mute: z.boolean().optional(),
+      loop: z.boolean().optional(),
+      start: z.preprocess(
+        (val) => {
+          if (val === null || val === undefined || val === '' || (typeof val === 'number' && isNaN(val))) return undefined;
+          return typeof val === 'string' ? parseFloat(val) : val;
+        },
+        z.number().optional()
+      ),
+      end: z.preprocess(
+        (val) => {
+          if (val === null || val === undefined || val === '' || (typeof val === 'number' && isNaN(val))) return undefined;
+          return typeof val === 'string' ? parseFloat(val) : val;
+        },
+        z.number().optional()
+      ),
+      clickToLoad: z.boolean().optional(),
+      showMuteButton: z.boolean().optional(),
+      videoOnly: z.boolean().optional(),
+      svgatorSync: z.boolean().optional(),
+      svgatorId: z.string().optional(),
+      interactiveMode: z.enum(['sync', 'independent', 'controller']).optional(),
     }).optional()
   }).optional(),
 });
 
-const contactPage = defineCollection({
-  type: 'data',
-  schema: z.object({
-    content: z.string().optional(),
-    showName: z.boolean().optional(),
-    showPhone: z.boolean().optional(),
-    showMessage: z.boolean().optional(),
-    showUpload: z.boolean().optional(),
-    extraFieldLabel: z.string().optional(),
-    showExtraField: z.boolean().optional(),
-    showMap: z.boolean().optional(),
-    extraFieldLabel2: z.string().optional(),
-    showExtraField2: z.boolean().optional(),
-  }),
-});
-
-const collections = {
-  posts: defineCollection({
+export const collections = {
+  // Post collection
+  post: defineCollection({
     type: 'content',
     schema: postSchema,
   }),
 
-  contactPage,
+  // StyleApps collection
+  styleapps: defineCollection({
+    type: 'data',
+    schema: z.object({
+      backgroundImage: z.string().optional(),
+      backgroundVideo: z.string().optional(),
+      siteFont: z.string().optional(),
+      borderRadius: z.string().optional(),
+      lightBg: z.string().optional(),
+      lightAccent2: z.string().optional(),
+      darkBg: z.string().optional(),
+      darkAccent2: z.string().optional(),
+      lightHeader: z.string().optional(),
+      darkHeader: z.string().optional(),
+      lightCardBg: z.string().optional(),
+      darkCardBg: z.string().optional(),
+      lightText: z.string().optional(),
+      darkText: z.string().optional(),
+      customCSS: z.string().optional()
+    })
+  }),
 
   pages: defineCollection({
     type: 'content',
     schema: z.object({
-      title: z.string().optional(),
+      title: z.string(),
+      slug: z.string().optional(),
       description: z.string().optional(),
-      content: z.string().optional(),
+      layout: z.string().optional(),
+      draft: z.boolean().optional(),
+      // New template system - allows pages to use homepage components
+      useTemplateSystem: z.boolean().optional(),
+      sections: z.array(z.object({
+        type: z.enum([
+          'contentblock', 'pitch', 'testimonials', 'faqs', 
+          'resume', 'ctas', 'youtubefeeds', 'youform', 'posts', 'magicsearch', 'photos', 'app', 'form'
+        ]),
+        customTitle: z.string().optional(),
+        customDescription: z.string().optional(),
+        showTitle: z.boolean().optional(),
+        sectionWidth: z.enum(['narrow', 'normal', 'wide', 'full']).optional(),
+        // Content block specific
+        contentBlockSlug: z.string().optional(), // Reference to pitch slug
+        // YouTube feed specific
+        feedConfig: z.string().optional(), // Reference to YouTube feed slug
+        // CTA specific
+        cta: z.string().optional(), // Reference to CTA slug
+        showSearch: z.boolean().optional(),
+        searchMethod: z.enum(['client', 'pagefind', 'hybrid']).optional(),
+        hideCollapseButton: z.boolean().optional(),
+      })).optional().default([]),
     }),
   }),
-
-  CTAs: defineCollection({
-    type: 'data',
-    schema: z.object({
-      title: z.string().optional(),
-      ctaUrl: z.string().optional(),
-      description: z.string().optional(),
-      showFancy: z.boolean().optional(),
-      showTransition: z.boolean().optional(),
-    }),
-  }),
-
-  
 
   pitches: defineCollection({
-    type: 'data',
-    schema: z.object({
-      title: z.string().optional(),
-      showTwocol: z.boolean().optional(),
-      showTitle: z.boolean().optional(),
-      image: z.string().optional(),
-      imageAlt: z.string().optional(),
-      description: z.string().optional(),
-      tagline: z.string().optional(),
-      subheading1: z.string().optional(),
-      text1: z.string().optional(),
-      subheading2: z.string().optional(),
-      text2: z.string().optional(),
-      subheading3: z.string().optional(),
-      text3: z.string().optional(),
-    }),
+    type: 'content',
+    schema: z.any(), // Flexible schema to match Keystatic
   }),
 
   faqs: defineCollection({
@@ -115,7 +185,7 @@ const collections = {
   resume: defineCollection({
     type: 'content',
     schema: z.object({
-      section: z.string(),
+      section: z.string().optional(),
       showTitle: z.boolean().optional(),
       content: z.string().optional(),
     }),
@@ -132,29 +202,25 @@ const collections = {
     }),
   }),
 
+
+
+  // Optional menuItems collection - can be empty if not needed
   menuItems: defineCollection({
     type: 'data',
     schema: z.object({
+      name: z.string().optional(),
       title: z.string().optional(),
       path: z.string().optional(),
       order: z.number().optional(),
     }),
   }),
 
-  piratePosts: defineCollection({
-    type: 'content',
-    schema: z.object({
-      title: z.string().optional(),
-      content: z.string().optional(),
-      createdAt: z.string().or(z.date()).transform((val) => new Date(val)),
-    }),
-  }),
-
-  pirateFeeds: defineCollection({
+  footerMenuItems: defineCollection({
     type: 'data',
     schema: z.object({
+      name: z.string().optional(),
       title: z.string().optional(),
-      feedUrl: z.string().optional(),
+      path: z.string().optional(),
       order: z.number().optional(),
     }),
   }),
@@ -165,7 +231,7 @@ const collections = {
       friendlyName: z.string().optional(),
       link: z.string().optional(),
       icon: z.string().optional(),
-      isWebmention: z.boolean().optional(),
+      isActive: z.boolean().optional(),
       order: z.any().transform(val => 
         (val === '.nan' || val === 'nan' || Number.isNaN(val)) ? undefined : Number(val)
       ).optional()
@@ -178,11 +244,10 @@ const collections = {
       logoImage: z.string().optional(),
       showHeader: z.boolean().optional(),
       showLogo: z.boolean().optional(),
-      showCheck: z.boolean().optional(),
       showHome: z.boolean().optional(),
       showTheme: z.boolean().optional(),
       showSwitch: z.boolean().optional(),
-      showSearch: z.boolean().optional(),
+      // showSearch: z.boolean().optional(),
       showFooter: z.boolean().optional(),
       defaultView: z.enum(['grid', 'swipe']).optional(),
       themeMode: z.enum(['light', 'dark', 'user']).optional(),
@@ -190,10 +255,12 @@ const collections = {
       showDates: z.boolean().optional(),
       enableImageBlur: z.boolean().optional(),
       showTags: z.boolean().optional(),
+      showTagFilters: z.boolean().optional(),
       showSocial: z.boolean().optional(),
       MAX_POSTS: z.number().optional(),
       MAX_POSTS_PER_PAGE: z.number().optional(),
       showShare: z.boolean().optional(),
+      videoTimeLimitMinutes: z.number().min(-1).max(30).optional(),
     }),
   }),
 
@@ -205,6 +272,7 @@ const collections = {
       name: z.string().optional(),
       shortName: z.string().optional(),
       screenshot: z.string().optional(),
+      title: z.string().optional(),
       description: z.string().optional(),
       themeColor: z.string().optional(),
       backgroundColor: z.string().optional(),
@@ -213,59 +281,17 @@ const collections = {
       icon192: z.string().optional(),
       icon512: z.string().optional(),
       location: z.string().optional(),
-    }),
-  }),
-
-  home: defineCollection({
-    type: 'data',
-    schema: z.object({
-      showFeature: z.boolean().optional(),
-      featureImage: z.object({
-        src: z.string().optional(),
-        alt: z.string().optional(),
-      }).optional(),
-      youtube: z.object({
-        discriminant: z.boolean(),
-        value: z.object({
-          url: z.string().optional(),
-          title: z.string().optional(),
-          controls: z.boolean().optional(),
-          useCustomPlayer: z.boolean().optional(),
-          mute: z.boolean().optional(),
-          loop: z.boolean().optional(),
-          start: z.number().optional(),
-          end: z.number().optional(),
-          videoOnly: z.boolean().optional(),
-        }).optional()
-      }).optional(),
-      cta: z.string().optional(),
-      showBioOnHome: z.boolean().optional(),
-      showApp: z.boolean().optional(),
-      showHomeGallery: z.boolean().optional(),
-      showResume: z.boolean().optional(),
-      showPosts: z.boolean().optional(),
-      showMore: z.boolean().optional(),
-      showFaqOnHome: z.boolean().optional(),
-      showTestOnHome: z.boolean().optional(),
-      pitch: z.string().optional(),
-      pitch2: z.string().optional(),
-      pitch3: z.string().optional(),
-      featureOrder: z.number().optional(),
-      bioOrder: z.number().optional(),
-      appOrder: z.number().optional(),
-      galleryOrder: z.number().optional(),
-      postsOrder: z.number().optional(),
-      resumeOrder: z.number().optional(),
-      faqOrder: z.number().optional(),
-      testimonialsOrder: z.number().optional(),
-      infoblockOrder: z.number().optional(),
-      infoblock2Order: z.number().optional(),
-      infoblock3Order: z.number().optional(),
-      photosectiontitle: z.string().optional(),
-      locationtitle: z.string().optional(),
-      faqsectiontitle: z.string().optional(),
-      testimonialtitle: z.string().optional(),
-      postsectiontitle: z.string().optional(),
+      showMap: z.boolean().optional(),
+      // Contact form configuration
+      showName: z.boolean().optional(),
+      showPhone: z.boolean().optional(),
+      showMessage: z.boolean().optional(),
+      showUpload: z.boolean().optional(),
+      showExtraField: z.boolean().optional(),
+      extraFieldLabel: z.string().optional(),
+      showExtraField2: z.boolean().optional(),
+      extraFieldLabel2: z.string().optional(),
+      formContent: z.string().optional(),
     }),
   }),
 
@@ -274,39 +300,16 @@ const collections = {
     schema: z.object({
       galleryMode: z.enum(['directory', 'keystatic']).optional(),
       showCaptions: z.boolean().optional(),
-      showBioOnPhotos: z.boolean().optional(),
-      showFaqsOnPhotos: z.boolean().optional(),
-      showTestimonialsOnPhotos: z.boolean().optional(),
-      pitch: z.string().optional(),
+      autoOpenLightbox: z.boolean().optional(),
+      // showFaqsOnPhotos: z.boolean().optional(),
+      // showTestimonialsOnPhotos: z.boolean().optional(),
+      // pitch: z.string().optional(),
       defaultDirectory: z.string().optional(),
       showGallerySelector: z.boolean().optional(),
       galleryImages: z.array(z.object({
         image: z.string().optional(),
         caption: z.string().optional(),
       })).optional(),
-    }),
-  }),
-
-  styleAppearance: defineCollection({
-    type: 'data',
-    schema: z.object({
-      backgroundImage: z.string().optional(),
-      backgroundVideo: z.string().optional(),
-      siteFont: z.string().optional(),
-      borderRadius: z.string().optional(),
-      lightBg: z.string().optional(),
-      lightAccent: z.string().optional(),
-      lightAccent2: z.string().optional(),
-      darkBg: z.string().optional(),
-      darkAccent: z.string().optional(),
-      darkAccent2: z.string().optional(),
-      lightHeader: z.string().optional(),
-      darkHeader: z.string().optional(),
-      lightText: z.string().optional(),
-      darkText: z.string().optional(),
-      lightLink: z.string().optional(),
-      darkLink: z.string().optional(),
-      customCSS: z.string().optional(),
     }),
   }),
 
@@ -320,7 +323,7 @@ const collections = {
       viewmore: z.string().optional(),
       allimages: z.string().optional(),
       close: z.string().optional(),
-      search: z.string().optional(),
+      // search: z.string().optional(),
       mute: z.string().optional(),
       volume: z.string().optional(),
       progress: z.string().optional(),
@@ -329,34 +332,14 @@ const collections = {
       shareText: z.string().optional(),
       copyButton: z.string().optional(),
       siteDisclaimer: z.string().optional(),
+      socialMessage: z.string().optional(),
+
     }),
   }),
 
-  bio: defineCollection({
-    type: 'data',
-    schema: z.object({
-      title: z.string().optional(),
-      tagline: z.string().optional(),
-      description: z.string().optional(),
-      image: z.string().optional(),
-      name: z.string().optional(),
-      phone: z.string().optional(),
-      subheading: z.string().optional(),
-      subcontent: z.string().optional(),
-      cta: z.string().optional(),
-      showSocial: z.boolean().optional(),
-    }),
-  }),
 
-  pirateSocial: defineCollection({
-    type: 'data',
-    schema: z.object({
-      profile: z.string().optional(),
-      description: z.string().optional(),
-      autoDeletePiratePosts: z.boolean().optional(),
-      autoDeleteTime: z.number().optional(),
-    }),
-  }),
+
+
 
   resumeSettings: defineCollection({
     type: 'data',
@@ -369,24 +352,55 @@ const collections = {
       rightColumnItems: z.array(z.string()).optional(),
     }),
   }),
-};
 
-export { collections };
+  CTAs: defineCollection({
+    type: 'data',
+    schema: z.object({
+      title: z.string().optional(),
+      ctaUrl: z.string().optional(),
+      description: z.string().optional(),
+      showFancy: z.boolean().optional(),
+      showTransition: z.boolean().optional()
+    })
+  }),
 
-export type PitchData = {
-  slug: string;
-  showTitle: boolean | null;
-  tagline: string;
-  text1: string;
-  text2: string;
-  text3: string;
-  subheading1: string;
-  subheading2: string;
-  subheading3: string;
-  image: string | null;
-  imageAlt: string;
-  description: string;
-  title: string | null;
-  divider: string | null;
-  divider2: string | null;
+  // youtubeFeedCollections: defineCollection({
+  //   type: 'data',
+  //   schema: z.object({
+  //     name: z.string(),
+  //     description: z.string().optional(),
+  //     channels: z.array(z.object({
+  //       channelId: z.string(),
+  //       channelName: z.string().optional()
+  //     })).default([]),
+  //     category: z.enum(['tech', 'education', 'science', 'entertainment', 'news', 'politics', 'podcasts', 'gaming', 'lifestyle']).default('tech')
+  //   })
+  // }),
+
+  youtubeFeeds: defineCollection({
+    type: 'data',
+    schema: z.object({
+      title: z.string().optional(),
+      description: z.string().optional(),
+      feedType: z.enum(['channels', 'collections', 'mixed']).default('channels'),
+      channelIds: z.array(z.string()).default([]),
+      selectedCollections: z.array(z.string()).default([]),
+      maxVideos: z.number().min(1).max(250).default(6),
+      initialLoad: z.number().min(1).max(100).default(21),
+      showTitles: z.boolean().default(true),
+      showSectionTitle: z.boolean().default(true),
+      useCustomPlayer: z.boolean().default(false),
+      defaultView: z.enum(['grid', 'swipe']).default('grid'),
+      includeSitePosts: z.boolean().default(false),
+      maxSitePosts: z.number().min(0).max(20).default(3)
+    })
+  }),
+
+  // Social Card collection for default site image (from photoUpload directory)
+  photoUpload: defineCollection({
+    type: 'data',
+    schema: z.object({
+      socialCard: z.string().optional(), // Image path from Keystatic upload
+    })
+  }),
 };
